+++ src/main/java/util/dataProvider/SystemDefaultZoneCurrentDateProvider.java
@@ -1,0 +1,11 @@
+package util.dataProvider;
+
+import java.time.LocalDate;
+
+public class SystemDefaultZoneCurrentDateProvider implements DateProvider {
+    
+    @Override
+    public LocalDate getDate() {
+        return LocalDate.now();
+    }
+}
