--- src/main/java/util/dataTimeProvider/DefaultDateProvider.java
@@ -1,11 +1,0 @@
-package util.dataTimeProvider;
-
-import java.time.LocalDate;
-
-public class DefaultDateProvider implements DateProvider {
-    
-    @Override
-    public LocalDate getCurrentDate() {
-        return LocalDate.now();
-    }
-}
